Refactored transaction
